extended: move keystore var, DHT struct to relevant files
