+++ extended.go
@@ -2,6 +2,17 @@ package rtfs
 
 import "context"
 
+// DHTFindProvsResponse is a response from the findprovs command
+type DHTFindProvsResponse struct {
+	ID        string `json:"id,omitempty"`
+	Type      int    `json:"type,omitempty"`
+	Responses [][]struct {
+		ID    string   `json:"id,omitempty"`
+		Addrs []string `json:"addrs,omitempty"`
+	} `json:"responses,omitempty"`
+	Extra string `json:"extra,omitempty"`
+}
+
 // DHTFindProvs is used to find providers of a given CID
 // Currently bugged and wil only fetch 1 provider
 func DHTFindProvs(im Manager, cid, numProviders string) error {
